[lib] refactor logic for csv validation
- To have all the csv checks in one file (csv_validator.go).
